Added Ralma
Ralma provides helper shortcuts for most Bulma widgets. The goal of this plugin is to help you get started quickly and reduce the overly verbose code that Bulma tends to produce. It does not attempt to cover everything in Bulma, but rather offer the most common elements as shortcuts. Sometimes, when you want something more specific or fancy, you can still use the original Bulma markup.